Rename tests directory
